pyproject.toml
[project]
name = "practical-dl-at-scale-with-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.23",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "mlflow>=2.22.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.7.1",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "transformers>=4.52.3",
]
